Move and update isYes helper function
