+++ src/applications/survivors-benefits/utils/helpers.jsx
@@ -172,3 +172,9 @@ export async function addStyleToShadowDomOnPages(
       }
     });
 }
+
+// helper to handle both boolean and string "yes" values coming from different
+// test fixtures or UX components
+export const isYes = val =>
+  val === true ||
+  (typeof val === 'string' && val.toLowerCase().startsWith('y'));
